support copy-on-get and test client & host srcipts

+++ skystore_cli.py
@@ -84,12 +84,13 @@ def init(
         )
 
     # Start the skystore server
-    subprocess.Popen(
-        f"cd {DEFAULT_STORE_SERVER_PATH}; "
-        "rm skystore.db; python3 -m uvicorn app:app --reload --port 3000",
-        shell=True,
-        env=env,
-    )
+    if start_server:
+        subprocess.Popen(
+            f"cd {DEFAULT_STORE_SERVER_PATH}; "
+            "rm skystore.db; python3 -m uvicorn app:app --reload --port 3000",
+            shell=True,
+            env=env,
+        )
 
     time.sleep(2)
 
@@ -120,7 +121,8 @@ def register(
     if local_test:
         server_addr = "localhost"
     else:
-        server_addr = "15.160.154.191"
+        # NOTE: ip address set to be the remote store-server addr
+        server_addr = "54.183.123.82"
 
     try:
         with open(register_config, "r") as f:
